feat:마이페이지 찜관리 퍼블리싱완료
마이페이지 찜관리 퍼블리싱 완료했고, url 정리 했습니다.
