+++ src/app/(house-keeper)/[id]/page.tsx
@@ -551,7 +551,7 @@ const ListingStayDetailPage: FC<ListingStayDetailPageProps> = ({}) => {
         </div>
 
         {/* SUBMIT */}
-        <ButtonPrimary href={"/checkout"}>Reserve</ButtonPrimary>
+        <ButtonPrimary href={"/serviceform"}>Reserve</ButtonPrimary>
       </div>
     );
   };
