+++ src/app/(listing-detail)/(components)/ModalReserveMobile.tsx
@@ -1,5 +1,5 @@
 import React, { FC, Fragment, useState } from "react";
-import CheckOutPagePageMain from "@/app/checkout/PageMain";
+import CheckOutPagePageMain from "@/app/serviceform/PageMain";
 import { Dialog, Transition } from "@headlessui/react";
 import { XMarkIcon } from "@heroicons/react/24/solid";
