+++ src/app/(listing-detail)/listing-experiences-detail/page.tsx
@@ -418,7 +418,7 @@ const ListingExperiencesDetailPage: FC<
         </div>
 
         {/* SUBMIT */}
-        <ButtonPrimary href={"/checkout"}>Reserve</ButtonPrimary>
+        <ButtonPrimary href={"/serviceform"}>Reserve</ButtonPrimary>
       </div>
     );
   };
