+++ src/app/(review)/review/[id]/page.tsx
@@ -1,4 +1,4 @@
-import CheckOutPagePageMain from "@/app/checkout/PageMain";
+
 import React from "react";
 import ReviewDetail from "./PageMain";
